Create Code-AccesPoint-Mode
Added alternative coding where the wifi credentials do not have to be hardcoded. If the ESP cannot connect to a network it will turn into an wifi access point itself where when you connect to it with your phone and go to its IP adress you should be able to enter and save your wifi credentials dynamically.